### 2024-10-04 #663 @willgearty
rredlist: 'IUCN' Red List Client
Submitting Author Name: William Gearty
Submitting Author Github Handle: @willgearty
Other Package Authors Github handles: (comma separated, delete if none) @sckott, @maelle
Repository:  https://github.com/ropensci/rredlist
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: rredlist
Title: 'IUCN' Red List Client
Description: 'IUCN' Red List (<https://api.iucnredlist.org/>) client.
    The 'IUCN' Red List is a global list of threatened and endangered species.
    Functions cover all of the Red List 'API' routes. An 'API' key is required.
Version: 0.7.1.9000
    c(person(given = "William",
             family = "Gearty",
             email = "willgearty@gmail.com"),
      person(given = "Scott",
             family = "Chamberlain",
             role = c("aut"),
             email = "myrmecocystus@gmail.com"),
      person(given = "rOpenSci",
             role = "fnd",
             comment = "https://ropensci.org/"),
      person(given = "Maëlle",
             family = "Salmon",
             role = "ctb",
             email = "maelle.salmon@yahoo.se")
      )
     https://github.com/ropensci/rredlist
    crul (>= 0.3.8),
    jsonlite (>= 1.1),
    lifecycle
    vcr (>= 0.4.0),
X-schema.org-applicationCategory: Biodiversity
X-schema.org-keywords: IUCN, biodiversity, API, web-services, traits, habitat, species, conservation
X-schema.org-isPartOf: https://ropensci.org
```

Package retrieves data from the IUCN Red List.

The IUCN Red List is used by a broad range of users, including biologists, ecologists, conservationists, activists, etc. It has many uses including

There are no other packages, to my knowledge, that perform the same function.

N/A

N/A

The `pkgcheck` package says my package only has 11.2% coverage, but [my own coverage checks](https://app.codecov.io/github/ropensci/rredlist?branch=master) calculate the coverage to be ~94%.

Package is already on CRAN.
