### 2024-10-07 #663 @willgearty
I should mention that `rredlist` was originally developed by @sckott and is already a part of ropensci. However, the IUCN Red List API has recently changed dramatically, so I basically rewrote the entire package to work with the new API. I thought now was as good a time as any to have the package finally go through ropensci review (with the intent to submit to JOSS afterwards).
