### 2024-10-23 #663 @robitalec
Oh @willgearty, one more thing. I missed this in the [editor checks](https://github.com/ropensci/software-review/issues/663#issuecomment-2394102292) under Statistical Properties. I see doclines_per_fn_not_exp ("median number of lines of documentation for each non-exported R function") is 0. Looking through your R/ directory, I found these non-exported R functions without any documentation.

release.R:
- `release_questions`

zzz.R:
- `ct`
- `rredlist_ua`
- `rr_GET`
- `err_catcher`
- `rl_parse`
- `check_key`
- `rr_base`
- `space`
- `assert_is`
- `assert_n`
- `assert_not_na`
- `combine_assessments`
- `page_assessments`

I would suggest adding some internal documentation for anyone trying to debug an issue when using {rredlist}, or to help potential contributors understand how {rredlist} works. You can use the `#' @noRd` tag ([roxygen2 help](https://roxygen2.r-lib.org/reference/tags-rd.html), [dev guide](https://devguide.ropensci.org/pkg_building.html#roxygen-2-use)) to prevent .Rd generation. I would think eg. `rr_GET`, `rr_base`, `rredlist_ua` could be good candidates for brief documentation given they are quite central to the package ([network of calls](https://ropensci-review-tools.github.io/roreviewapi/static/rredlist_pkgstats4fd92778.html)). Thanks!
