### 2024-10-23 caf5f65
Remove sapply; update README
