### 2024-10-23 1622204
Add overview of functions to vignette
