+++ vignettes/source/_rredlist.Rmd
@@ -97,6 +97,17 @@ rredlist::rl_use_iucn()
 
 Note that you can no longer generate a new API key for version 3 of the API.
 
+## Overview of available functions
+
+- Query assessments by taxa (e.g., `rl_species()` and `rl_family()`)
+- Query assessments by habitats (e.g., `rl_habitats()` and `rl_systems()`)
+- Query assessments by geography (e.g., `rl_countries()` and `rl_realms()`)
+- Query assessments by actions needed (e.g., `rl_actions()` and `rl_research()`)
+- Query assessments by threat details (e.g., `rl_threats()` and `rl_pop_trends()`)
+- Query assessments for specific sets of taxa (e.g., `rl_extinct()` and `rl_growth_forms()`)
+- Get all details for a single assessment (e.g., `rl_species_latest()` and `rl_assessment()`)
+- Red List information and statistics (e.g., `rl_sp_count()` and `rl_version()`)
+
 ## High level interface
 
 ```{r}
