### 2024-10-23 30464bf
vcr updates, including docs in CONTRIBUTING.md
