+++ .github/CONTRIBUTING.md
@@ -1,17 +1,0 @@
-# CONTRIBUTING #
-
-### Bugs?
-
-* Submit an issue on the [Issues page](https://github.com/ropensci/rredlist/issues)
-
-### Code contributions
-
-* Fork this repo to your Github account
-* Clone your version on your account down to your machine from your account, e.g,. `git clone https://github.com/<yourgithubusername>/rredlist.git`
-* Make sure to track progress upstream (i.e., on our version of `rredlist` at `ropensci/rredlist`) by doing `git remote add upstream https://github.com/ropensci/rredlist.git`. Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step
-* Make your changes (bonus points for making changes on a new feature branch)
-* Run tests!
-* Push up to your account
-* Submit a pull request to home base at `ropensci/rredlist`
-
-### Also, check out our [discussion forum](https://discuss.ropensci.org)
