### 2024-10-23 #663 @willgearty
@robitalec thanks for flagging those issues. I believe I have addressed all of your concerns in preparation for review!
