### 2024-10-29 #663 @robitalec
Thanks to both @stephhazlitt and @KevCaz for agreeing to review {rredlist}!

Please let me know if you need anything as you work on your review.
