### 2024-11-08 1dfd910
fix typos and replace \code{} by markdown tags
