+++ man-roxygen/curl.R
@@ -1,2 +1,2 @@
 #' @param ... [Curl options][curl::curl_options()] passed to the GET request via
-#'   \code{\link[crul]{HttpClient}}.
+#'   [cruls][crul::HttpClient()].
