+++ man-roxygen/all.R
@@ -1,3 +1,3 @@
-#' @param key (character) An IUCN API token. See \code{\link{rl_use_iucn}}.
-#' @param parse (logical) Whether to parse the output to list (\code{FALSE}) or,
-#'   where possible, data.frame (\code{TRUE}). Default: \code{TRUE}
+#' @param key (character) An IUCN API token. See [rl_use_iucn()].
+#' @param parse (logical) Whether to parse the output to list (`FALSE`) or,
+#'   where possible, data.frame (`TRUE`). Default: `TRUE`.
