### 2024-11-09 #663 @robitalec
Thank you @KevCaz for your review.

Just a minor note, I see under "Could have" / "tests" two paragraphs with a few repeated sentences. When you have a moment, can you double check and possibly edit to ensure your comments about tests are clear for everyone?
