### 2024-11-11 99dc079
fix typos and replace \code{} by markdown tags (#60)
