+++ man-roxygen/commonargs.R
@@ -1,5 +1,5 @@
-#' @param name (character) A taxonomic name
-#' @param id (character) An IUCN identifier
-#' @param region (character) A region name, see \code{\link{rl_regions}} for
-#' acceptable region identifiers (use the entries in the \code{identifier}
-#' column)
+#' @param name (character) A taxonomic name.
+#' @param id (character) An IUCN identifier.
+#' @param region (character) A region name, see [rl_regions()] for
+#' acceptable region identifiers (use the entries in the `identifier`
+#' column).
