### 2024-11-14 #663 @maelle
So it was a glitch, sorry @KevCaz you are now properly registered!
