### 2024-11-20 #663 @stephhazlitt
## `rredlist` Package Review (@stephhazlitt)

- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work.

Estimated hours spent reviewing: ~4

---

### Review Comments:

This is a nicely focused, tidy and well-documented data retrieval package that closely follows the rOpenSci packaging guide. In addition, the author has already implemented some incremental improvements from the rOpenSci editor (@robitalec) and community (@KevCaz) reviews. Apart from a few minor code-related comments, I have focused my review on my user experience and thinking of how analysts and researchers might interact with this software. I had fun playing with this package 🦋🐝🦆🌿! Thanks for the opportunity to review, I hope the comments are useful.

_General Comments_

- Saving the API Key: I think for new users (and novice R users) a smooth path with the API key will be the "key" (pun intended). The `rredlist::rl_use_iucn()` helper fxn to get and save the key takes you to the web page and prints the instructions to save to the console&mdash;I missed the latter because I was whisked-away+distracted with the IUCN web page. I think it would a nice enhancement to mature this workflow with a fxn that wraps `usethis` and helps a user get the key sorted either in their `.Rprofile` (my usual workflow) or `.Renviron` (I gather their is no consensus re: a [recommended approach](https://www.r-bloggers.com/2015/11/how-to-store-and-use-webservice-keys-and-authentication-details-with-r)). It might also be nice to alert users in the README/Vignette that they need to provide an email and other business information to the IUCN.
- First Argument Options: I anticipate users often needing to look-up the first argument options for the various query fxns. It is mentioned in the `.Rd` docs that calling the fxn with no first argument does this for you (e.g., `If not supplied, a list of all kingdom names  will be returned.`), however this could also be exampled in the README+vignette. I suspect this might be a common, interactive part of a typical workflow?

```
rl_kingdom()
$kingdom_names
[1] "ANIMALIA"  "CHROMISTA" "FUNGI"     "PLANTAE"
```
- An Even Tidier API?: I appreciate the different APIs, higher-level vs low-level. I wonder if a future enhancement could push the higher-level API to follow an even "tidier" approach. A list object, in my experience, is less data analysis-ready than anything data frame. Since the package already parses the most data-rich element of the list to a data frame, I wonder about taking it a bit further and adding the other list objects as metadata columns and returning a data frame. This would perhaps result in 3 APIs: a tidy one (returning a data frame), a list object (currently `parse = FALSE`) and the low level API (returning just the JSON). In the short-term, it might be useful to add some hints in the vignette about how to navigate the list objects (e.g., `names(list)` )?
- Progress Bar: I think it would be useful to add in a progress bar for users in place of the `.`s&mdash;no matter the API used, the larger downloads take time, it is nice to have something to stare at.
- Speed: The vignette mentions speed being one reason to leverage the lower-level API. Although not at all thorough, I did not find the low-level API significantly faster on the couple of large data pulls I tried.

```
tic()
ex1 <- rl_class(class = "Aves")
..........................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................
toc()
1420.385 sec elapsed

tic()
ex2 <- rl_class_(class = "Aves")
..........................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................
toc()
1267.238 sec elapsed
```

It might be worth sharing some benchmarking so users know where the time is used (download vs parse) and how or if they can further optimize data retrieval speed. For me, setting `parse = FALSE` felt like it saved a little time up front, but most of the time was the data transfer itself?

_Minor Comments_
- It would be nice to see the production branch renamed following current community norms (`master` -> `main`)
- The `rl_sp_count()` returns an integer not a list (need to override `#' @template info` for this one)
- Would it be useful to embed a check-for-the-internet fxn and throw an informative error if not connected?
- I am a bird nerd by training and immediately went to pull all the assessments for the Oystercatchers of the world. Is there a path to allow a user to pull all the records for a given Genus? (e.g., `rl_species(genus = "Haematopus")`)
- I wonder about shifting the API versioning explanation to the end of the vignette, and then consolidate the sentence about not being able to get an API key for the older version?
