### 2024-12-02 #663 @willgearty
Thanks for the check-in @robitalec. With Thanksgiving last week, I haven't had a chance to work on any changes to the package, but I have read over the reviews and everything sounds reasonable and I do plan to implement nearly all of the suggestions hopefully over the next week.
