### 2024-12-04 50a1bf2
Add cli progress bar and page assessments using response headers
