+++ NAMESPACE
@@ -84,6 +84,9 @@ export(rl_use_and_trade)
 export(rl_use_and_trade_)
 export(rl_use_iucn)
 export(rl_version)
+importFrom(cli,cli_progress_bar)
+importFrom(cli,cli_progress_done)
+importFrom(cli,cli_progress_update)
 importFrom(crul,HttpClient)
 importFrom(jsonlite,fromJSON)
 importFrom(lifecycle,deprecated)
