### 2024-12-05 c315fd1
Check for internet before each GET call
