+++ NAMESPACE
@@ -88,6 +88,7 @@ importFrom(cli,cli_progress_bar)
 importFrom(cli,cli_progress_done)
 importFrom(cli,cli_progress_update)
 importFrom(crul,HttpClient)
+importFrom(curl,nslookup)
 importFrom(jsonlite,fromJSON)
 importFrom(lifecycle,deprecated)
 importFrom(utils,browseURL)
