### 2024-12-05 8a688aa
Catch server error status codes and throw friendly error
