### 2024-12-05 b162c31
Remove err_catcher() to reduce download overhead
