### 2024-12-05 dc1011e
Add docs for space()
