### 2024-12-05 718845b
More informative progress bar
