### 2024-12-05 f20c52a
Add test coverage for other status codes
