### 2024-12-05 75bdf5a
Remove unused assert function
