### 2024-12-06 5b99b75
Add tests for utility functions
