+++ tests/fixtures/Rheidae_metadata.yml
@@ -1,0 +1,39 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://api.iucnredlist.org/api/v4/taxa/family/Rheidae?page=1
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      User-Agent: r-curl/6.0.1 crul/1.5.0 rOpenSci(rredlist/0.7.1.9000)
+      Accept-Encoding: gzip, deflate
+      Accept: application/json, text/xml, application/xml, */*
+      Authorization: My bearer token is safe
+  response:
+    status:
+      status_code: '200'
+      message: OK
+      explanation: Request fulfilled, document follows
+    headers:
+      status: HTTP/1.1 200 OK
+      cache-control: max-age=0, private, must-revalidate
+      content-length: '3149'
+      content-type: application/json
+      current-page: '1'
+      etag: W/"ffa325e8d9058c13de669a118541da23"
+      link: <https://api.iucnredlist.org/api/v4/taxa/family/Rheidae?page=1&per_page=100>;
+        rel="first", <https://api.iucnredlist.org/api/v4/taxa/family/Rheidae?page=1&per_page=100>;
+        rel="last"
+      page-items: '100'
+      total-count: '15'
+      total-pages: '1'
+      x-request-id: aff56963-4314-4eb8-8ecc-d06f5ef5ccc2
+      x-runtime: '0.525640'
+      date: Fri, 06 Dec 2024 21:11:41 GMT
+    body:
+      encoding: ''
+      file: no
+      string: '{"assessments":[{"year_published":"2008","latest":false,"sis_taxon_id":22678073,"url":"https://www.iucnredlist.org/species/22678073/34712085","assessment_id":34712085,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"2004","latest":false,"sis_taxon_id":22678073,"url":"https://www.iucnredlist.org/species/22678073/34713259","assessment_id":34713259,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"2000","latest":false,"sis_taxon_id":22678073,"url":"https://www.iucnredlist.org/species/22678073/34714240","assessment_id":34714240,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"1994","latest":false,"sis_taxon_id":22678073,"url":"https://www.iucnredlist.org/species/22678073/34714746","assessment_id":34714746,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"1988","latest":false,"sis_taxon_id":22678073,"url":"https://www.iucnredlist.org/species/22678073/34715731","assessment_id":34715731,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"2012","latest":false,"sis_taxon_id":22678073,"url":"https://www.iucnredlist.org/species/22678073/37835094","assessment_id":37835094,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"2014","latest":false,"sis_taxon_id":22728199,"url":"https://www.iucnredlist.org/species/22728199/40855535","assessment_id":40855535,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"2014","latest":false,"sis_taxon_id":22728206,"url":"https://www.iucnredlist.org/species/22728206/40856094","assessment_id":40856094,"scopes":[{"description":{"en":"Global"},"code":"1"}]},{"year_published":"2016","latest":false,"sis_taxon_id":22678073,"url":"https://www.iucnredlist.org/species/22678073/92754472","assessment_id":92754472,"scopes":[{"description":{"en":"Global"},"
[…]
