+++ tests/fixtures/rr_GET_raw.yml
@@ -1,0 +1,34 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://api.iucnredlist.org/api/v4/taxa/family
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      User-Agent: r-curl/6.0.1 crul/1.5.0 rOpenSci(rredlist/0.7.1.9000)
+      Accept-Encoding: gzip, deflate
+      Accept: application/json, text/xml, application/xml, */*
+      Authorization: My bearer token is safe
+  response:
+    status:
+      status_code: '200'
+      message: OK
+      explanation: Request fulfilled, document follows
+    headers:
+      status: HTTP/1.1 200 OK
+      cache-control: max-age=0, private, must-revalidate
+      content-length: '41748'
+      content-type: application/json
+      etag: W/"7c0f81a478719d9fd124b511da922ffe"
+      x-request-id: fa3575e1-cb00-4625-b765-0ad202fea0f9
+      x-runtime: '0.014924'
+      date: Fri, 06 Dec 2024 16:34:47 GMT
+    body:
+      encoding: ''
+      file: no
+      string: '{"family_names":["ABROCOMIDAE","ACANTHACEAE","ACANTHAMETROPODIDAE","ACANTHISITTIDAE","ACANTHIZIDAE","ACANTHODRILIDAE","ACANTHOGORGIIDAE","ACANTHURIDAE","ACAROSPORACEAE","ACAVIDAE","ACCIPITRIDAE","ACESTRORHYNCHIDAE","ACHARIACEAE","ACHATINELLIDAE","ACHATINIDAE","ACHATOCARPACEAE","ACHEILOGNATHIDAE","ACHIPTERIIDAE","ACHIRIDAE","ACHIROPSETTIDAE","ACICULIDAE","ACIPENSERIDAE","ACORACEAE","ACRIDIDAE","ACROBATIDAE","ACROBOLBACEAE","ACROCEPHALIDAE","ACROCERIDAE","ACROCHAETIACEAE","ACROCHORDIDAE","ACROLOXIDAE","ACROPOMATIDAE","ACROPORIDAE","ACTINIDIACEAE","ACTINIIDAE","ADELANTHACEAE","ADRIANICHTHYIDAE","AEGITHALIDAE","AEGITHINIDAE","AEGOTHELIDAE","AENIGMACHANNIDAE","AENIGMATINEIDAE","AESHNIDAE","AETOBATIDAE","AEXTOXICACEAE","AFROTHISMIACEAE","AGAMIDAE","AGARDHIELLIDAE","AGARICACEAE","AGARICIIDAE","AGELENIDAE","AGONIDAE","AGRIOLIMACIDAE","AGROMYZIDAE","AILIIDAE","AILURIDAE","AIPTASIIDAE","AIZOACEAE","AKANIACEAE","AKYSIDAE","ALARIACEAE","ALAUDIDAE","ALBATRELLACEAE","ALBULIDAE","ALCEDINIDAE","ALCIDAE","ALCIPPEIDAE","ALCYONII
[…]
