### 2024-12-06 4a40ffb
Clean up rredlist-package.R
