### 2024-12-09 0eb9b1e
Return bibentry for rl_citation()
