+++ NAMESPACE
@@ -91,5 +91,6 @@ importFrom(crul,HttpClient)
 importFrom(curl,nslookup)
 importFrom(jsonlite,fromJSON)
 importFrom(lifecycle,deprecated)
+importFrom(utils,bibentry)
 importFrom(utils,browseURL)
 importFrom(utils,packageVersion)
