### 2024-12-10 0d4c2bf
Fix rl_citation tests
