+++ tests/testthat/test-rl_citation.R
@@ -7,9 +7,10 @@ test_that("rl_citation works", {
     aa <- rl_citation()
   })
 
-  expect_is(aa, "character")
-  expect_match(aa, "IUCN")
-  expect_match(aa, "www.iucnredlist.org")
+  expect_is(aa, "bibentry")
+  aa_print <- paste(capture.output(aa), collapse = " ")
+  expect_match(aa_print, "IUCN Red List of Threatened Species")
+  expect_match(aa_print, "www.iucnredlist.org")
 })
 
 test_that("rl_citation curl options work", {
