+++ NAMESPACE
@@ -84,6 +84,11 @@ export(rl_use_and_trade)
 export(rl_use_and_trade_)
 export(rl_use_iucn)
 export(rl_version)
+importFrom(cli,cli_code)
+importFrom(cli,cli_end)
+importFrom(cli,cli_h2)
+importFrom(cli,cli_li)
+importFrom(cli,cli_ol)
 importFrom(cli,cli_progress_bar)
 importFrom(cli,cli_progress_done)
 importFrom(cli,cli_progress_update)
@@ -92,5 +97,4 @@ importFrom(curl,nslookup)
 importFrom(jsonlite,fromJSON)
 importFrom(lifecycle,deprecated)
 importFrom(utils,bibentry)
-importFrom(utils,browseURL)
 importFrom(utils,packageVersion)
