### 2024-12-10 a6cbfe3
Fix docs for rl_sp_count()
