### 2024-12-11 bbc6fd2
Add benchmarking vignette
