+++ vignettes/benchmarks.Rmd
@@ -1,0 +1,199 @@
+---
+# Generated by vignettes/precompile.R: do not edit by hand.
+# Please edit source in vignettes/source/_benchmarks.Rmd
+title: "rredlist benchmarks"
+author: "William Gearty"
+date: "2024-12-10"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{rredlist benchmarks}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+
+## Introduction
+
+`rredlist` provides two APIs, a higher-level one that takes slightly more time but returns the data in a more user-friendly format (a list), and a lower-level one (i.e., functions that end with "_") that takes less time but does no processing of the data (returning the raw JSON string). Both APIs return the exact same information, but it is up to the user whether the format processing is worth the extra time, especially when performing bulk operations. To help inform this decision by the user, here is some benchmarking related to the two APIs. First, we'll break down the total difference in computation time between the two APIs, then we'll dig into what components are causing this difference. We'll use `microbenchmark::microbenchmark()` which has very little computational overhead. Note that the time units vary from comparison to comparison, and the speed of these functions may be highly hardware- and network-dependent.
+
+
+``` r
+library(rredlist)
+library(microbenchmark)
+```
+
+## Head-to-head benchmarks
+
+We'll start by benchmarking the two APIs head-to-head. We'll test a couple of use cases, in rough order of increasing complexity.
+
+#### 1\. Get species count
+
+
+``` r
+microbenchmark(
+  rl_sp_count(),
+  rl_sp_count_(),
+  times = 10
+)
+#> Unit: milliseconds
+#>            expr      min       lq     mean   median       uq      max neval cld
+#>   rl_sp_count() 112.1753 112.9675 116.0770 114.6603 119.2189 126.5307    10   a
+#>  rl_sp_count_() 111.1262 112.3976 116.1929 115.4099 116.0395 127.9950    10   a
+```
+
+#### 2\. Lookup individual assessment
+
+
+``` r
+microbenchmark(
+  rl_assessment(136250858),
+  rl_assessment_(136250858),
+  times = 10
+)
+#> Unit: milliseconds
+#>                       expr      min       lq     mean   median       uq      max neval cld
+#>   rl_assessment(136250858) 257.1335 266.4728 286.9152 277.9695 313.7179 322.6544    10   a
+#>  rl_assessment_(136250858) 252.2234 260.1433 311.6918 287.6650 323.1625 478.7026    10   a
+```
+
+#### 3\. Taxonomic lookup with defaults
+
+
+``` r
+microbenchmark(
+  rl_family(),
+  rl_family_(),
+  times = 10
+)
+#> Unit: milliseconds
+#>          expr      min       lq     mean   median       uq      max neval cld
+#>   rl_family() 118.1405 119.3698 123.6975 120.1616 123.0535 144.8783    10   a
+#>  rl_family_() 116.6910 118.6265 156.5042 119.3431 132.2629 428.2468    10   a
+```
+
+#### 4\. Taxonomic lookup with query (one page of results)
+
+
+``` r
+microbenchmark(
+  rl_family("Rheidae"),
+  rl_family_("Rheidae"),
+  times = 10
+)
+#> Unit: milliseconds
+#>                   expr      min       lq     mean   median       uq      max neval cld
+#>   rl_family("Rheidae") 616.2717 617.7706 626.2586 621.4997 637.6830 647.0362    10   a
[…]
