+++ vignettes/source/_benchmarks.Rmd
@@ -1,0 +1,179 @@
+---
+title: "rredlist benchmarks"
+author: "William Gearty"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{rredlist benchmarks}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r echo=FALSE}
+library("knitr")
+library("rredlist")
+hook_output <- knitr::knit_hooks$get("output")
+knitr::knit_hooks$set(output = function(x, options) {
+   lines <- options$output.lines
+   if (is.null(lines)) {
+     return(hook_output(x, options))  # pass to default hook
+   }
+   x <- unlist(strsplit(x, "\n"))
+   more <- "..."
+   if (length(lines)==1) {        # first n lines
+     if (length(x) > lines) {
+       # truncate the output, but add ....
+       x <- c(head(x, lines), more)
+     }
+   } else {
+     x <- c(if (abs(lines[1])>1) more else NULL,
+            x[lines],
+            if (length(x)>lines[abs(length(lines))]) more else NULL
+           )
+   }
+   # paste these lines together
+   x <- paste(c(x, ""), collapse = "\n")
+   hook_output(x, options)
+ })
+
+knitr::opts_chunk$set(
+  warning = FALSE,
+  message = FALSE,
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## Introduction
+
+`rredlist` provides two APIs, a higher-level one that takes slightly more time but returns the data in a more user-friendly format (a list), and a lower-level one (i.e., functions that end with "_") that takes less time but does no processing of the data (returning the raw JSON string). Both APIs return the exact same information, but it is up to the user whether the format processing is worth the extra time, especially when performing bulk operations. To help inform this decision by the user, here is some benchmarking related to the two APIs. First, we'll break down the total difference in computation time between the two APIs, then we'll dig into what components are causing this difference. We'll use `microbenchmark::microbenchmark()` which has very little computational overhead. Note that the time units vary from comparison to comparison, and the speed of these functions may be highly hardware- and network-dependent.
+
+```{r}
+library(rredlist)
+library(microbenchmark)
+```
+
+## Head-to-head benchmarks
+
+We'll start by benchmarking the two APIs head-to-head. We'll test a couple of use cases, in rough order of increasing complexity.
+
+#### 1\. Get species count
+
+```{r}
+microbenchmark(
+  rl_sp_count(),
+  rl_sp_count_(),
+  times = 10
+)
+```
+
+#### 2\. Lookup individual assessment
+
+```{r}
+microbenchmark(
+  rl_assessment(136250858),
+  rl_assessment_(136250858),
+  times = 10
+)
+```
+
+#### 3\. Taxonomic lookup with defaults
+
+```{r}
+microbenchmark(
+  rl_family(),
+  rl_family_(),
+  times = 10
+)
+```
+
+#### 4\. Taxonomic lookup with query (one page of results)
+
+```{r}
+microbenchmark(
+  rl_family("Rheidae"),
+  rl_family_("Rheidae"),
+  times = 10
+)
+```
+
+#### 5\. Taxonomic lookup with query (~11 pages of results)
+
+```{r}
+microbenchmark(
+  rl_family("Corvidae", quiet = TRUE),
+  rl_family_("Corvidae", quiet = TRUE),
+  times = 10
+)
+```
+
+#### 6\. Taxonomic lookup with query (~37 pages of results)
+
+```{r}
+microbenchmark(
[…]
