+++ vignettes/source/_benchmarks.Rmd
@@ -97,7 +97,7 @@ microbenchmark(
 )
 ```
 
-#### 5\. Taxonomic lookup with query (~11 pages of results)
+#### 5\. Taxonomic lookup with query (~10 pages of results)
 
 ```{r}
 microbenchmark(
@@ -107,7 +107,7 @@ microbenchmark(
 )
 ```
 
-#### 6\. Taxonomic lookup with query (~37 pages of results)
+#### 6\. Taxonomic lookup with query (~40 pages of results)
 
 ```{r}
 microbenchmark(
@@ -117,7 +117,7 @@ microbenchmark(
 )
 ```
 
-#### 7\. Taxonomic lookup with query (~922 pages of results)
+#### 7\. Taxonomic lookup with query (~900 pages of results)
 
 ```{r}
 microbenchmark(
@@ -171,7 +171,7 @@ microbenchmark(
 )
 ```
 
-Again, even with 37 pages of data to parse, the download takes the vast majority of the time. The highest-level parsing has increased to about 1% of the time it takes to download the data, but this remains less than a second compared to the ~35 second download.
+Again, even with about 40 pages of data to parse, the download takes the vast majority of the time. The highest-level parsing has increased to about 1% of the time it takes to download the data, but this remains less than a second compared to the ~35 second download.
 
 ## Conclusion
 
