### 2024-12-11 2d91ac2
Move API versioning docs to end of vignette; clean up vignette
