+++ vignettes/rredlist.Rmd
@@ -3,7 +3,7 @@
 # Please edit source in vignettes/source/_rredlist.Rmd
 title: "Introduction to rredlist"
 author: "William Gearty"
-date: "2024-10-23"
+date: "2024-12-11"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Introduction to rredlist}
@@ -13,10 +13,10 @@ vignette: >
 
 
 
-`rredlist` is an R client for the IUCN Red List API (https://api.iucnredlist.org/).
+`rredlist` is an R client for the [IUCN Red List API](https://api.iucnredlist.org/).
 The [IUCN Red List](https://www.iucnredlist.org/) is the world’s most comprehensive information source on the global extinction risk status of animal, fungus and plant species. This package provides access via R to the various data contained within this database which span range details, population size, habitat and ecology, use and/or trade, threats, and conservation actions.
 
-Documentation for the IUCN Red List API are available at https://api.iucnredlist.org/api-docs. **Note that use of the web API, and therefore this package, [requires authentication](#authentication).**
+Documentation for the IUCN Red List API is available at https://api.iucnredlist.org/api-docs. **Note that use of the web API, and therefore this package, [requires authentication](#authentication).**
 
 #### What `rredlist` is not:
 
@@ -43,23 +43,9 @@ remotes::install_github("ropensci/rredlist")
 install.packages("rredlist", repos = "https://ropensci.r-universe.dev/")
 ```
 
-### API Versioning
-
-`rredlist` versions ≤ 0.7.1 tracked [version 3](https://apiv3.iucnredlist.org/) of the IUCN Red List API. `rredlist` versions ≥ 1.0.0 track version 4 of the IUCN Red List API. If you need to use version 3 of the API--and it is still functioning--you can install an old version of `rredlist` using the `remotes` package:
-
-
-``` r
-# From CRAN archive
-remotes::install_version("rredlist", version = "0.7.1")
-
-# From r-universe
-remotes::install_version("rredlist", version = "0.7.1",
-                         repos = "https://ropensci.r-universe.dev/")
-```
-
 ## Authentication
 
-IUCN requires you to get your own API key, an alphanumeric string that you need to send in every request. There's a helper function in the package helping you getting it at https://api.iucnredlist.org/users/sign_up and storing it. Once you store this key, the functions in `rredlist` will always default to using this key.
+IUCN requires you to get your own API key, an alphanumeric string that you need to send in every request. The rl_use_iucn() function in the package can walk you through [getting your own key](https://api.iucnredlist.org/users/sign_up) and storing it locally. Once you store this key, the functions in `rredlist` will always default to using this key.
 
 
 ``` r
@@ -67,8 +53,6 @@ rredlist::rl_use_iucn()
 ```
 
 **Keep this key private.** You can pass the key in to each function via the `key` parameter, but it's better to store the key either as a environment variable (`IUCN_REDLIST_KEY`) or an R option (`iucn_redlist_key`) - we recommend using the former option.
-
[…]
