### 2024-12-11 6a3e003
Make rl_citation() print BibTeX entry automatically
