+++ man/rl_sis.Rd
@@ -18,7 +18,7 @@ rl_sis_(id, key = NULL, ...)
 where possible, data.frame (\code{TRUE}). Default: \code{TRUE}.}
 
 \item{...}{\link[curl:curl_options]{Curl options} passed to the GET request via
-\link[crul:HttpClient]{cruls}.}
+\link[crul:HttpClient]{crul}.}
 }
 \value{
 A list unless using a function with a trailing underscore, in which
