+++ man/rl_sis_latest.Rd
@@ -15,7 +15,7 @@ rl_sis_latest(id, key = NULL, parse = TRUE, ...)
 where possible, data.frame (\code{TRUE}). Default: \code{TRUE}.}
 
 \item{...}{\link[curl:curl_options]{Curl options} passed to the GET request via
-\link[crul:HttpClient]{cruls}.}
+\link[crul:HttpClient]{crul}.}
 }
 \value{
 A list unless using a function with a trailing underscore, in which
