### 2024-12-12 a76f939
Switch to HttpClient()$retry() and document related arguments
