### 2024-12-13 47da382
Assert that queries have a length of 1
