+++ R/rl_categories.R
@@ -41,6 +41,7 @@ rl_categories_ <- function(code = NULL, key = NULL, all = TRUE, page = 1,
                            quiet = FALSE, ...) {
   assert_is(key, "character")
   assert_is(code, "character")
+  assert_n(code, 1)
   assert_is(page, c("integer", "numeric"))
   assert_n(page, 1)
   assert_is(all, "logical")
