+++ R/rl_habitats.R
@@ -37,6 +37,7 @@ rl_habitats_ <- function(code = NULL, key = NULL, all = TRUE, page = 1,
                          quiet = FALSE, ...) {
   assert_is(key, "character")
   assert_is(code, "character")
+  assert_n(code, 1)
   assert_is(page, c("integer", "numeric"))
   assert_n(page, 1)
   assert_is(all, "logical")
@@ -89,6 +90,7 @@ rl_systems_ <- function(code = NULL, key = NULL, all = TRUE, page = 1,
                         quiet = FALSE, ...) {
   assert_is(key, "character")
   assert_is(code, "character")
+  assert_n(code, 1)
   assert_is(page, c("integer", "numeric"))
   assert_n(page, 1)
   assert_is(all, "logical")
