+++ R/rl_realms.R
@@ -36,6 +36,7 @@ rl_realms_ <- function(code = NULL, key = NULL, all = TRUE, page = 1,
                        quiet = FALSE, ...) {
   assert_is(key, "character")
   assert_is(code, "character")
+  assert_n(code, 1)
   assert_is(page, c("integer", "numeric"))
   assert_n(page, 1)
   assert_is(all, "logical")
@@ -89,6 +90,7 @@ rl_scopes_ <- function(code = NULL, key = NULL, all = TRUE, page = 1,
                        quiet = FALSE, ...) {
   assert_is(key, "character")
   assert_is(code, "character")
+  assert_n(code, 1)
   assert_is(page, c("integer", "numeric"))
   assert_n(page, 1)
   assert_is(all, "logical")
@@ -143,6 +145,7 @@ rl_faos_ <- function(code = NULL, key = NULL, all = TRUE, page = 1,
                      quiet = FALSE, ...) {
   assert_is(key, "character")
   assert_is(code, "character")
+  assert_n(code, 1)
   assert_is(page, c("integer", "numeric"))
   assert_n(page, 1)
   assert_is(all, "logical")
