### 2024-12-14 626e70f
Add ggplot2 scale for IUCN color palette
