+++ NAMESPACE
@@ -84,6 +84,10 @@ export(rl_use_and_trade)
 export(rl_use_and_trade_)
 export(rl_use_iucn)
 export(rl_version)
+export(scale_color_iucn)
+export(scale_colour_iucn)
+export(scale_discrete_iucn)
+export(scale_fill_iucn)
 importFrom(cli,cli_code)
 importFrom(cli,cli_end)
 importFrom(cli,cli_h2)
@@ -96,5 +100,6 @@ importFrom(crul,HttpClient)
 importFrom(curl,nslookup)
 importFrom(jsonlite,fromJSON)
 importFrom(lifecycle,deprecated)
+importFrom(rlang,check_installed)
 importFrom(utils,bibentry)
 importFrom(utils,packageVersion)
