### 2024-12-15 056650e
Add scale_color_iucn to pkgdown index
