### 2024-12-15 c1f0b1d
Minor pkgdown tweaks
