### 2024-12-15 c23f57d
Add research workflows vignette; improve precompile script
