+++ vignettes/research_workflows.Rmd
@@ -1,0 +1,492 @@
+---
+# Generated by vignettes/precompile.R: do not edit by hand.
+# Please edit source in vignettes/source/_research_workflows.Rmd
+title: "Using rredlist within a research workflow"
+author: "William Gearty"
+date: "2024-12-14"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Using rredlist within a research workflow}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+<div style="text-align: justify">
+
+## Introduction
+
+Once you have `rredlist` [set up](rredlist.html), it's very likely you'll want to use it for some sort of research workflow. This vignette gives examples of the kinds of data that users may be interested in downloading from the IUCN API, and how they might then use that data within a partial research workflow. I've given three worked examples, but because the possibilities are pretty endless, I've also cited some recent published works at the bottom of the vignette to cover a wider variety of use cases.
+
+Before we get started, we'll just load some packages that we'll need. We'll use `{dplyr}` for handling and cleaning our data, and we'll use `{ggplot2}` and `{patchwork}` for visualizing our data.
+
+
+``` r
+library(rredlist)
+library(dplyr)
+library(ggplot2)
+library(patchwork)
+```
+
+## How have mollusc assessments increased through time?
+
+For our first research workflow, we are interested in the assessment of molluscs through time. Specifically, we are interested in how many cumulative assessments have been completed each year, and how many cumulative species have then been assessed each year. First, we'll need to download an assessment summary for Mollusca. Note that while I've used `quiet = TRUE` here for the purposes of the vignette, you would likely want to skip that to get a lovely progress bar that keeps you apprised of the download's progress since there are quite a few pages of assessments to download.
+
+
+``` r
+moll_assessments <- rl_phylum(phylum = "Mollusca", quiet = TRUE)
+```
+
+Now that we have our assessment summary, we can summarize the data by calculating the cumulative numbers of assessments through time. This is a fairly basic `{dplyr}` pipeline that counts the number of assessments for each year, then orders those counts from oldest to most recent, then calculates the cumulative sum of the counts.
+
+
+``` r
+moll_counts <- moll_assessments$assessments %>%
+  group_by(year_published) %>%
+  summarize(assess_cnt = n()) %>%
+  ungroup() %>%
+  mutate(year_published = as.numeric(year_published)) %>%
+  arrange(year_published) %>%
+  mutate(assess_cum = cumsum(assess_cnt))
+
+```
+
+Now that we have our cumulative sums, we can make a simple plot. We'll save it for later to combine with our second plot.
+
+
+``` r
+gg1 <- ggplot(moll_counts) +
+  geom_line(aes(x = year_published, y = assess_cum)) +
+  labs(x = "Year", y = "Cumulative # of Assessments") +
+  theme_classic(base_size = 14)
+```
+
+Now let's see how many cumulative _species_ are assessed by year. In order to do this, we'll first need to filter to only the first assessment for each species using `arrange()` and `slice()`. Then we can calculate the cumulative sum as we did for the assessments above.
+
+
+``` r
+moll_sp_counts <- moll_assessments$assessments %>%
+  mutate(year_published = as.numeric(year_published)) %>%
+  group_by(sis_taxon_id) %>%
+  arrange(year_published) %>%
+  slice(1) %>%
+  ungroup() %>%
+  group_by(year_published) %>%
+  summarize(sp_cnt = n()) %>%
+  ungroup() %>%
+  mutate(year_published = as.numeric(year_published)) %>%
[…]
