### 2024-12-15 126ad82
Improve other vignettes
