+++ vignettes/rredlist.Rmd
@@ -3,7 +3,7 @@
 # Please edit source in vignettes/source/_rredlist.Rmd
 title: "Introduction to rredlist"
 author: "William Gearty"
-date: "2024-12-11"
+date: "2024-12-15"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Introduction to rredlist}
@@ -54,7 +54,7 @@ rredlist::rl_use_iucn()
 
 **Keep this key private.** You can pass the key in to each function via the `key` parameter, but it's better to store the key either as a environment variable (`IUCN_REDLIST_KEY`) or an R option (`iucn_redlist_key`) - we recommend using the former option.
 
-## Overview of available functions
+## Overview of available features
 
 - Query assessments by taxa (e.g., `rl_species()` and `rl_family()`)
 - Query assessments by habitats (e.g., `rl_habitats()` and `rl_systems()`)
@@ -65,7 +65,74 @@ rredlist::rl_use_iucn()
 - Get all details for a single assessment (e.g., `rl_species_latest()` and `rl_assessment()`)
 - Red List information and statistics (e.g., `rl_sp_count()` and `rl_version()`)
 
-## High level interface
+## Example usage
+
+### Loading the package
+
+``` r
+library("rredlist")
+```
+
+### Search for assessments for a particular species
+
+``` r
+rl_species("Gorilla", "gorilla")$assessments
+#>    year_published latest possibly_extinct possibly_extinct_in_the_wild sis_taxon_id
+#> 1            2016  FALSE            FALSE                        FALSE         9404
+#> 2            2018   TRUE            FALSE                        FALSE         9404
+#> 3            2016  FALSE            FALSE                        FALSE         9404
+#> 4            2008  FALSE            FALSE                        FALSE         9404
+#> 5            2007  FALSE            FALSE                        FALSE         9404
+#> 6            2000  FALSE            FALSE                        FALSE         9404
+#> 7            1996  FALSE            FALSE                        FALSE         9404
+#> 8            1994  FALSE            FALSE                        FALSE         9404
+#> 9            1990  FALSE            FALSE                        FALSE         9404
+...
+```
+
+### Search for assessments that recommend particular conservation actions
+
+#### Get a list of all conservation actions
+
+``` r
+rl_actions()
+#> $conservation_actions
+#>                                              en  code
+#> 1                         Land/water protection     1
+#> 2                          Site/area protection   1_1
+#> 3                 Resource & habitat protection   1_2
+#> 4                         Land/water management     2
+#> 5                          Site/area management   2_1
+#> 6          Invasive/problematic species control   2_2
+#> 7         Habitat & natural process restoration   2_3
+#> 8                            Species management     3
+...
+```
+
+#### Return assessments with a particular conservation action
+
+``` r
+rl_actions("2_2", all = FALSE)$assessments
+#>     year_published latest possibly_extinct possibly_extinct_in_the_wild sis_taxon_id
+#> 1             2019   TRUE            FALSE                        FALSE    132523146
[…]
