+++ vignettes/source/_rredlist.Rmd
@@ -82,7 +82,7 @@ rredlist::rl_use_iucn()
 
 **Keep this key private.** You can pass the key in to each function via the `key` parameter, but it's better to store the key either as a environment variable (`IUCN_REDLIST_KEY`) or an R option (`iucn_redlist_key`) - we recommend using the former option.
 
-## Overview of available functions
+## Overview of available features
 
 - Query assessments by taxa (e.g., `rl_species()` and `rl_family()`)
 - Query assessments by habitats (e.g., `rl_habitats()` and `rl_systems()`)
@@ -93,7 +93,37 @@ rredlist::rl_use_iucn()
 - Get all details for a single assessment (e.g., `rl_species_latest()` and `rl_assessment()`)
 - Red List information and statistics (e.g., `rl_sp_count()` and `rl_version()`)
 
-## High level interface
+## Example usage
+
+### Loading the package
+```{r}
+library("rredlist")
+```
+
+### Search for assessments for a particular species
+```{r output.lines=1:10}
+rl_species("Gorilla", "gorilla")$assessments
+```
+
+### Search for assessments that recommend particular conservation actions
+
+#### Get a list of all conservation actions
+```{r output.lines=1:10}
+rl_actions()
+```
+
+#### Return assessments with a particular conservation action
+```{r output.lines=1:10}
+rl_actions("2_2", all = FALSE)$assessments
+```
+
+## Advanced usage
+
+For a more in-depth walkthrough of how to integrate `rredlist` into research workflows, check out [the vignette](research_workflows.html).
+
+## High level vs. low level interfaces
+
+### High level interface
 
 ```{r}
 library("rredlist")
@@ -121,7 +151,7 @@ rl_species("Fratercula", "arctica", parse = FALSE)$assessments
 
 For even more speed, you can use the low level package interface (although see the [benchmarking vignette](benchmarks.html) for specifics).
 
-## Low level interface
+### Low level interface
 
 The parsing to data.frame in the high level functions does take extra time. The low level functions
 only do the HTTP request, and give back JSON without doing any more parsing. The low level functions DO have an underscore on the end of the function name, e.g., `rl_species_()`:
