### 2024-12-15 bcb4b6c
Add tests for scales
