### 2024-12-15 #663 @willgearty
Apologies for the delay, but I've now addressed nearly all of the review comments in this PR: https://github.com/ropensci/rredlist/pull/62. I've summarized the changes below. A huge thank you again to the reviewers who both provided extremely useful suggestions, and please don't hesitate to let me know if you have any other comments!

## @KevCaz's comments:

- Progress bar: very useful suggestion; added in [50a1bf2](https://github.com/ropensci/rredlist/pull/62/commits/50a1bf2a9959c39f8cf503cec874514c64f5f006) and improved in [718845b](https://github.com/ropensci/rredlist/pull/62/commits/718845b6c2b83d030ed42b528584fb9876b89566)
  - I looked into moving page_assessments within rr_GET, but I'm not sure the benefits would be worth the effort
- Second vignette: this was a great suggestion, and I've added it in [c23f57d](https://github.com/ropensci/rredlist/pull/62/commits/c23f57d8cfdd3e469bbd05ee457d90acf6dba3a1)
- test: I've added tests for the utility functions as suggested in [5b99b75](https://github.com/ropensci/rredlist/pull/62/commits/5b99b7560d1a80264c3ec2b675164713fbcfc5d6)
- vignette overview of functions: I've renamed this section as suggested, but I'm not entirely sure what a table would look like for this. Let me know if you have more detail about what you were thinking here.
- rl_citation(): this was a great suggestion, the output of rl_citation() now includes the accession date and a BibTex entry ([0eb9b1e](https://github.com/ropensci/rredlist/pull/62/commits/0eb9b1ed11492e7ad35b37fe76acc88695edb005) and [6a3e003](https://github.com/ropensci/rredlist/pull/62/commits/6a3e00341569e27ebb03d374d83c07471ff95f80))
- rredlist-package.R: cleaned this up in [4a40ffb](https://github.com/ropensci/rredlist/pull/62/commits/4a40ffb411300c211ce2b5a3077c0f642185ee9a)
- README: I've added the suggested text in [64cb3b9](https://github.com/ropensci/rredlist/pull/62/commits/64cb3b92865d330b7bf34074c788cc4f3a6d8310)
- precompilation of vignettes: I prefer this setup because RStudio doesn't give the pretty styling of the markdown files if they aren't .Rmd files (e.g., if they end with `.origin`)

## @stephhazlitt's comments:

- rl_use_iucn(): thanks for the user insight on this! this function no longer opens the browser window automatically, instead providing step-by-step instructions for the user to follow (including linking to the URL for key signup and providing an example of how to edit the .Renviron file), implemented in [6e151ff](https://github.com/ropensci/rredlist/pull/62/commits/6e151ff3b72f702ce8dcbd06549a71848e000b28) and [1a3428e](https://github.com/ropensci/rredlist/pull/62/commits/1a3428e7862caa09bbf1c9362d4a8923c9857b52)
- first argument options: this is now mentioned in the new vignette ([c23f57d](https://github.com/ropensci/rredlist/pull/62/commits/c23f57d8cfdd3e469bbd05ee457d90acf6dba3a1))
- even tidier API: I looked into this, but I'm not sure there is much more that can be data.frame-ified. I agree data.frames are more user-friendly, but often the different list elements have different types of data, numbers of columns, etc. Please do let me know if there were particular parts of the returned objects that would be useful to have data.frame-d.
- progress bar: since I had only ever tested examples with a small number of pages, I never noticed how problematic this could be! progress bar implemented in [50a1bf2](https://github.com/ropensci/rredlist/pull/62/commits/50a1bf2a9959c39f8cf503cec874514c64f5f006) and improved in [718845b](https://github.com/ropensci/rredlist/pull/62/commits/718845b6c2b83d030ed42b528584fb9876b89566)
- speed: I've added a benchmark vignette in [bbc6fd2](https://github.com/ropensci/rredlist/pull/62/commits/bbc6fd2aaec727462ae9c23fc2dcf36aa90f8ed4); as you intuited, the vast majority of the time is spent downloading the data and parsing it takes relatively no time at all
- production branch renamed
- rl_sp_count(): docs updated in [a6cbfe3](https://github.com/ropensci/rredlist/pull/62/commits/a6cbfe3b1cf981f57a7cede4795c2db808c4f0be)
- internet check: I added a check for the internet in [c315fd1](https://github.com/ropensci/rredlist/pull/62/commits/c315fd1071da9d87c4d4bdf04748a19a4995a802)
- genus query: as far as I can tell, there is currently no way to lookup a genus in the API
- vignette layout: I've moved this part of the vignette to the end as suggested ([2d91ac2](https://github.com/ropensci/rredlist/pull/62/commits/2d91ac2768f13abca4af823e55451d5ae7f534c0))
