### 2024-12-20 #663 @robitalec
Thank you @willgearty for addressing the reviewer comments and your detailed response here.

@stephhazlitt and @KevCaz, please use the reviewer approval template available [here](https://devguide.ropensci.org/approval2template.html) to respond to @willgearty's changes. Please feel free to ask for any clarification or discuss suggestions further.

I understand many folks will be are on holidays around this time, so please let me know if you won't be able to continue the review until after the new year.

Thanks everyone!
