### 2024-12-23 #663 @stephhazlitt
## Reviewer Response: Steph Hazlitt

#### Final approval (post-review)

Thank you for the changes+improvements @willgearty, I am sure all of your additional documentation will really support the users of this data retrieval package.

- [x ] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 5 (total)

#### Comments

Sharing a couple of hiccups I encountered using the `ropensci-review` branch.

I struggled to get the new progress bar working, getting the below error message with `quiet = FALSE`. I tried with the package installed from my local copy and with a fresh install from GitHub (e.g., `remotes::install_github("ropensci/rredlist", ref = "ropensci-review")`).

```
ropensci-review* > rredlist::rl_phylum(phylum = "Mollusca", quiet = FALSE)
Error in `"id" %in% names(args)`:
! Could not evaluate cli `{}` expression: `pb_name`.
Caused by error in `eval(expr, envir = envir)`:
! object 'pb_name' not found
Type .Last.error to see the more details.
Error in `(function (e) …`:
! Error in a deferred `on.exit()` clause
Caused by error in `"id" %in% names(args)`:
! Could not evaluate cli `{}` expression: `pb_name`.
Caused by error in `eval(expr, envir = envir)`:
! object 'pb_name' not found
Type .Last.error to see the more details.
```

When I run `Check` locally I am getting a note:

```
❯ checking files in ‘vignettes’ ... NOTE
  The following directory looks like a leftover from 'knitr':
    ‘figure’
  Please remove from your package.

0 errors ✔ | 0 warnings ✔ | 1 note ✖

R CMD check succeeded
```

I added the below to my local `.Rbuildignore` file and the note went away (as suggested [here](https://stackoverflow.com/questions/40744735/leftovers-from-knitr)), in case helpful.

```
figure$
cache$
```
