### 2025-01-08 #663 @willgearty
Thanks @stephhazlitt, I'll address both of those issues ASAP.
