### 2025-01-12 835dee9
Fix progress bar and R-cmd-check
