+++ R/zzz.R
@@ -219,8 +219,8 @@ page_assessments <- function(path, key, quiet, ...) {
   if (length(total_pages) == 0) total_pages <- 1
   if (!quiet) cli_progress_bar(
     "Paging assessments", total = total_pages, clear = FALSE,
-    format = paste0("{pb_name} ({pb_current}/{pb_total}) | ",
-                    "{pb_bar} {pb_percent} | ETA: {pb_eta}"))
+    format = paste0("{cli::pb_name} ({cli::pb_current}/{cli::pb_total}) | ",
+                    "{cli::pb_bar} {cli::pb_percent} | ETA: {cli::pb_eta}"))
   tmp <- res$parse("UTF-8")
   if (total_pages == 1) {
     out <- tmp
