### 2025-01-12 #663 @willgearty
OK, I believe those issues should both be fixed now. @KevCaz please also let me know what you think of the updated package.
