### 2025-01-15 #663 @robitalec
(Apologies for my delay here!)

Thank you @stephhazlitt for your final approval and @willgearty for addressing these related issues. @KevCaz, please let us know your thoughts on {rredlist} after @willgearty's changes after the reviews.

Thanks everyone!
