+++ vignettes/rredlist.Rmd
@@ -3,7 +3,7 @@
 # Please edit source in vignettes/source/_rredlist.Rmd
 title: "Introduction to rredlist"
 author: "William Gearty"
-date: "2024-12-15"
+date: "2025-01-27"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Introduction to rredlist}
@@ -77,16 +77,16 @@ library("rredlist")
 
 ``` r
 rl_species("Gorilla", "gorilla")$assessments
-#>    year_published latest possibly_extinct possibly_extinct_in_the_wild sis_taxon_id
-#> 1            2016  FALSE            FALSE                        FALSE         9404
-#> 2            2018   TRUE            FALSE                        FALSE         9404
-#> 3            2016  FALSE            FALSE                        FALSE         9404
-#> 4            2008  FALSE            FALSE                        FALSE         9404
-#> 5            2007  FALSE            FALSE                        FALSE         9404
-#> 6            2000  FALSE            FALSE                        FALSE         9404
-#> 7            1996  FALSE            FALSE                        FALSE         9404
-#> 8            1994  FALSE            FALSE                        FALSE         9404
-#> 9            1990  FALSE            FALSE                        FALSE         9404
+#>    year_published latest possibly_extinct possibly_extinct_in_the_wild sis_taxon_id                                                url assessment_id    scopes
+#> 1            2016  FALSE            FALSE                        FALSE         9404 https://www.iucnredlist.org/species/9404/102330408     102330408 Global, 1
+#> 2            2018   TRUE            FALSE                        FALSE         9404 https://www.iucnredlist.org/species/9404/136250858     136250858 Global, 1
+#> 3            2016  FALSE            FALSE                        FALSE         9404  https://www.iucnredlist.org/species/9404/17963949      17963949 Global, 1
+#> 4            2008  FALSE            FALSE                        FALSE         9404  https://www.iucnredlist.org/species/9404/12983787      12983787 Global, 1
+#> 5            2007  FALSE            FALSE                        FALSE         9404  https://www.iucnredlist.org/species/9404/12983966      12983966 Global, 1
+#> 6            2000  FALSE            FALSE                        FALSE         9404  https://www.iucnredlist.org/species/9404/12983737      12983737 Global, 1
+#> 7            1996  FALSE            FALSE                        FALSE         9404  https://www.iucnredlist.org/species/9404/12983764      12983764 Global, 1
+#> 8            1994  FALSE            FALSE                        FALSE         9404  https://www.iucnredlist.org/species/9404/12984167      12984167 Global, 1
+#> 9            1990  FALSE            FALSE                        FALSE         9404  https://www.iucnredlist.org/species/9404/12984186      12984186 Global, 1
 ...
 ```
 
@@ -113,16 +113,16 @@ rl_actions()
 
 ``` r
 rl_actions("2_2", all = FALSE)$assessments
-#>     year_published latest possibly_extinct possibly_extinct_in_the_wild sis_taxon_id
-#> 1             2019   TRUE            FALSE                        FALSE    132523146
-#> 2             2019   TRUE            FALSE                        FALSE        10767
-#> 3             2013   TRUE            FALSE                        FALSE         1078
-#> 4             2019   TRUE            FALSE                        FALSE    132521900
-#> 5             2020  FALSE            FALSE                        FALSE         1086
-#> 6             2019   TRUE            FALSE                        FALSE         1117
-#> 7             2019   TRUE            FALSE                        FALSE        11797
-#> 8             2021   TRUE            FALSE                        FALSE        12124
[…]
