+++ vignettes/source/_research_workflows.Rmd
@@ -160,7 +160,7 @@ We could go ahead and download the full assessments for all Australian species n
 reptilia_assessments <- rl_class(class = "Reptilia", latest = TRUE, quiet = TRUE)
 ```
 
-OK, now we've got our assessment summaries for Autralian species and for reptiles. We can take those two lists of assessments and find the intersection of them to get the list of assessment IDs for Australian reptiles.
+OK, now we've got our assessment summaries for Australian species and for reptiles. We can take those two lists of assessments and find the intersection of them to get the list of assessment IDs for Australian reptiles.
 
 ```{r}
 assessment_ids <- intersect(au_assessments$assessments$assessment_id,
