### 2025-01-28 #663 @KevCaz
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 8h (total)

#### Comments

* Progress bar: looks great!
* Second vignette: great update! The new vignette covers exactly the sort of questions I had in mind, **very** useful for all users. I love the benchmark vignette too! :pray: One quick remark in the second vignette, you wrote *"For this download, we'll increase the number of retries to 10 (from the default of 3)."* I think it’s worth adding a sentence to explain this. My guess is that it’s a way to avoid redoing the large download in case of an issue.
* Thanks for testing the utils, by the way, it was super easy to track all your changes looking at the commit history, made this round of review very easy.
* "I've renamed this section as suggested", thanks, that's enough I think.
* Thanks for all the other changes!

I've found a couple of typos that I fixed in a PR (https://github.com/ropensci/rredlist/pull/65). Note that there were some changes in `rredlist.Rmd` as I rerun the compilation of all vignettes.

Thanks for all your hard work on this, and thanks for adding me as a reviewer, I changed my email and added my orcid in the second commit of the PR. Let me know if this works for you.
