### 2025-01-28 9722dcf
Fixing typos (#65)
* Fix typos

* Update Kevin Cazelles email and orcid
