+++ DESCRIPTION
@@ -24,7 +24,8 @@ Authors@R:
       person(given = "Kevin",
              family = "Cazelles",
              role = "rev",
-             email = "kevin.cazelles@gmail.com"),
+             email = "kevin.cazelles@insileco.io", 
+             comment = c(ORCID = "0000-0001-6619-9874")),
       person(given = "Stephanie",
              family = "Hazlitt",
              role = "rev",
