### 2025-01-28 4acade3
ropensci review changes (#62)
* Add cli progress bar and page assessments using response headers

* Check for internet before each GET call

* Catch server error status codes and throw friendly error

* Remove err_catcher() to reduce download overhead

* Add docs for space()

* More informative progress bar

* Add test coverage for other status codes

* Remove unused assert function

* Add tests for utility functions

* Clean up rredlist-package.R

* Add link to API docs to README

* Return bibentry for rl_citation()

* Fix rl_citation tests

* Improve user friendliness of rl_use_iucn()

* Cha
[…]
