### 2025-01-28 #663 @ropensci-review-bot
Approved! Thanks @willgearty for submitting and @stephhazlitt, @KevCaz for your reviews! :grin:
