### 2025-01-28 #663 @robitalec
Thanks to @stephhazlitt and @KevCaz for your reviews and @willgearty for responsive discussion and changes throughout.

Since the package was already under the ropensci namespace, I believe some of these todos won't be necessary - but I'll leave them to you @willgearty to double check on your side.
