### 2025-01-28 #663 @willgearty
Looks like I've covered all of those to-dos, thanks @robitalec for the smooth review process!

If I want to submit a paper to ME&E introducing the paper, is there anything I need to do on the ROpenSci side of things? Or do I just tell ME&E that I already went through ROpenSci review and link them here? (I realized I didn't check the box at the beginning, but I've changed my mind since)
