### 2025-01-29 #663 @robitalec
Thanks @willgearty, it was a pleasure helping with this review.

You can proceed directly to the MEE submission, and there will be a section during that process to indicate your submission has been reviewed and accepted by rOpenSci.

Here's a screenshot:
![Image](https://github.com/user-attachments/assets/700740d1-d298-435e-8887-0f2ccc651574)
